Fixes #22: and handle skip-tables parse defect
Apparently AWS returns storedBytes = 0 even when there are actually
logs in CloudWatch.  Now testing for the existence of
lastEventTimestamp to determine if a stream is empty or not.

While testing, discovered that there's some odd skip-grants entries
in the slow query log which break parsing.  Now we handle an empty
session as well as no longer failing when the parse fails.