Fixes #5: SSL Support added and no more 500s around cache concurrency
Added a lock so that the tied cache file won't be accessed
concurrently.  This is why we'd get 500 errors.

Also added SSL support.  using adhoc, so each time you start, you'll
need to get your browser to accept the self-signed cert. But it's a
start.

TODO:  instrument the lock around AWS calls and the cache.
Lots of copy/paste around lock management currently.